requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
redis==5.0.8
APScheduler==3.11.0
aliyun-oss2==2.18.4
gunicorn==23.0.0
gevent==24.2.1
gevent-websocket==0.10.1
python-dotenv==1.0.1
Werkzeug==3.0.3
